+++ packages/ui/src/index.tsx
@@ -5,8 +5,8 @@ import App from './view/App';
 import { ApiClient } from './api-client';
 import { createBearerTokenMiddleware } from './api-client/authentication/middleware';
 import { OAuth2AuthorisationCodeFlowTokenManager } from './api-client/authentication/OAuth2AuthorisationCodeFlowTokenManager';
-import { MemoryStorage } from './api-client/authentication/storage/MemoryStorage';
 import { getConfigFromEnv } from './config';
+import { WebStorage } from './api-client/authentication/storage/WebStorage';
 
 const config = getConfigFromEnv();
 
@@ -14,7 +14,7 @@ const tokenManager = new OAuth2AuthorisationCodeFlowTokenManager({
   clientId: config.cognitoClientId,
   handler: request => fetch(request),
   redirectUri: config.homepageUri,
-  storage: new MemoryStorage(),
+  storage: new WebStorage(localStorage),
   tokenEndpoint: `${config.cognitoApiUri}/oauth2/token`,
 });
 
@@ -27,7 +27,7 @@ const apiClient = new ApiClient({
 
 document.addEventListener('DOMContentLoaded', () => {
   ReactDOM.render(
-    <App apiClient={apiClient} />,
+    <App apiClient={apiClient} config={config} />,
     document.getElementById('app'),
   );
 });
