--- packages/ui/src/view/App.st.css
@@ -1,8 +1,0 @@
-:import {
-    -st-from: "./PuzzleInstructions.st.css";
-    -st-named: height as instructionsHeight;
-}
-
-.puzzle {
-    max-width: calc(100vmin - value(instructionsHeight));
-}
