+++ packages/ui/src/view/Puzzle.st.css
@@ -1,0 +1,8 @@
+:import {
+  -st-from: './PuzzleInstructions.st.css';
+  -st-named: height as instructionsHeight;
+}
+
+.root {
+  max-width: calc(100vmin - value(instructionsHeight));
+}
